move session struct and methods to session.go
